- updated app version to 2.0.2 - added dev server paybc proxy - added check before quick search to prevent errors - moved modals -> dialogs - cleaned up modals in App.vue - started refactoring UI component architecture (WIP but working atm) - moved some "not getters" - commented out apparently-not-used module
