+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "name-request",
-  "version": "2.0.1",
+  "version": "2.0.2",
   "private": true,
   "appName": "Name Request UI",
   "sbcName": "SBC Common Components",
