+++ src/components/common/index.ts
@@ -1,0 +1,1 @@
+export { default as AppTitleCols } from './app-title-cols.vue'
