+++ src/components/dialogs/index.ts
@@ -1,0 +1,17 @@
+export { default as AffiliationErrorDialog } from './affiliation-error.vue'
+export { default as CancelDialog } from './cancel.vue'
+export { default as ConditionsDialog } from './conditions.vue'
+export { default as ErrorDialog } from './error.vue'
+export { default as ExitDialog } from './exit.vue'
+export { default as HelpMeChooseDialog } from './help-me-choose.vue'
+export { default as LocationInfoDialog } from './location-info.vue'
+export { default as MrasSearchInfoDialog } from './mras-search-info.vue'
+export { default as NrNotRequiredDialog } from './nr-not-required.vue'
+export { default as PaymentDialog } from './payment.vue'
+export { default as PaymentCompleteDialog } from './payment-complete.vue'
+export { default as PickEntityOrConversionDialog } from './pick-entity-or-conversion.vue'
+export { default as PickRequestTypeDialog } from './pick-request-type.vue'
+export { default as ReapplyDialog } from './reapply.vue'
+export { default as ReceiptsDialog } from './receipts.vue'
+export { default as RefundDialog } from './refund.vue'
+export { default as UpgradeDialog } from './upgrade.vue'
