+++ src/components/dialogs/upgrade.vue
@@ -1,50 +1,31 @@
 <template>
-  <v-dialog max-width="45rem" :value="isVisible" persistent>
+  <v-dialog max-width="40%" :value="isVisible" persistent>
     <v-card>
 
       <v-card-title class="d-flex justify-space-between">
-        <div>Confirm Name Request</div>
-        <v-btn icon large class="dialog-close float-right" @click="hideModal()">
-          <v-icon>mdi-close</v-icon>
-        </v-btn>
+        <div>Upgrade to Priority</div>
       </v-card-title>
 
-      <v-card-text class="copy-normal pt-0">
-        <request-details
-          :applicant="applicant"
-          :name="name"
-          :nameChoices="nameChoices"
-        />
+      <v-card-text class="copy-normal">
         <fee-summary
           :filingData="[...paymentDetails]"
           :fees="[...paymentFees]"
         />
       </v-card-text>
 
-      <v-card-actions class="pt-4">
-        <v-btn v-if="allowCancel"
-               @click="cancelPayment()"
-               id="payment-cancel-btn"
-               class="button-red px-4"
-               text
-               :disabled="isLoadingPayment">Cancel Name Request</v-btn>
+      <v-card-actions>
         <v-spacer></v-spacer>
-        <v-btn @click="confirmPayment()"
-               id="payment-pay-btn"
-               class="primary px-4"
-               text
-               :loading="isLoadingPayment">Continue to Payment</v-btn>
-        <v-btn @click="hideModal()"
-               id="payment-close-btn"
-               class="button-blue"
-               :disabled="isLoadingPayment">Close</v-btn>
+        <v-btn @click="confirmPayment()" id="payment-pay-btn" class="primary" text :loading="isLoadingPayment">
+          Accept
+        </v-btn>
+        <v-btn @click="hideModal()" id="payment-close-btn" class="normal" text>Cancel</v-btn>
       </v-card-actions>
 
     </v-card>
   </v-dialog>
 </template>
 
-<script lang='ts'>
+<script lang="ts">
 import { Component, Mixins, Watch } from 'vue-property-decorator'
 import FeeSummary from '@/components/payment/fee-summary.vue'
 import RequestDetails from '@/components/common/request-details.vue'
@@ -55,48 +36,33 @@ import * as FilingTypes from '@/modules/payment/filing-types'
 import * as Jurisdictions from '@/modules/payment/jurisdictions'
 import { PaymentAction } from '@/enums'
 import { NameRequestMixin, PaymentMixin, PaymentSessionMixin, DisplayedComponentMixin } from '@/mixins'
-import { getBaseUrl } from './payment-utils'
+import { getBaseUrl } from '@/components/payment/payment-utils'
 
 @Component({
   components: {
     RequestDetails,
     FeeSummary
-  },
-  props: {
-    onActivate: {
-      type: Function,
-      default: async () => undefined
-    },
-    onCancel: {
-      type: Function,
-      default: async () => {}
-    }
   }
 })
-export default class PaymentModal extends Mixins(
+export default class UpgradeDialog extends Mixins(
   NameRequestMixin,
   PaymentMixin,
   PaymentSessionMixin,
   DisplayedComponentMixin
 ) {
   private isLoadingPayment: boolean = false
-
   /** The model value for the dialog component. */
   private isVisible = false
 
-  private get allowCancel (): boolean {
-    return (typeof this.$props.onCancel === 'function')
-  }
-
   /** Whether this modal should be shown (per store property). */
   private get showModal (): boolean {
-    return PaymentModule[PaymentTypes.PAYMENT_MODAL_IS_VISIBLE]
+    return PaymentModule[PaymentTypes.UPGRADE_MODAL_IS_VISIBLE]
   }
 
   /** Clears store property to hide this modal. */
   async hideModal () {
     this.isLoadingPayment = false
-    await PaymentModule.togglePaymentModal(false)
+    await PaymentModule.toggleUpgradeModal(false)
   }
 
   /** Depending on value, fetches fees and makes this modal visible or hides it. */
@@ -104,14 +70,9 @@ export default class PaymentModal extends Mixins(
   async onShowModal (val: boolean): Promise<void> {
     if (val) {
       const paymentConfig = {
-        filingType: FilingTypes.NM620,
+        filingType: FilingTypes.NM606,
         jurisdiction: Jurisdictions.BC,
         priorityRequest: this.priorityRequest || false
-      }
-
-      const { onActivate } = this.$props
-      if (typeof onActivate === 'function') {
-        onActivate(paymentConfig)
       }
 
       // only make visible on success, otherwise hide it
@@ -125,15 +86,14 @@ export default class PaymentModal extends Mixins(
     }
   }
 
-  /** Called when user clicks "Continue to Payment" button. */
+  /** Called when user clicks "Accept" button. */
   private async confirmPayment () {
     this.isLoadingPayment = true
     const { nrId, priorityRequest } = this
-
     const onSuccess = (paymentResponse) => {
       const { paymentId, paymentToken } = this
       // Save response to session
-      this.savePaymentResponseToSession(PaymentAction.CREATE, paymentResponse)
+      this.savePaymentResponseToSession(PaymentAction.UPGRADE, paymentResponse)
       // see if redirect is needed else go to existing NR screen
       const baseUrl = getBaseUrl()
       const redirectUrl = encodeURIComponent(`${baseUrl}/nr/${nrId}/?paymentId=${paymentId}`)
@@ -145,9 +105,9 @@ export default class PaymentModal extends Mixins(
     }
 
     const success = await this.createPayment({
-      action: PaymentAction.CREATE,
+      action: PaymentAction.UPGRADE,
       nrId: nrId,
-      filingType: FilingTypes.NM620,
+      filingType: FilingTypes.NM606,
       priorityRequest: priorityRequest
     } as CreatePaymentParams, onSuccess)
 
@@ -157,25 +117,15 @@ export default class PaymentModal extends Mixins(
     }
   }
 
-  /** Called when user clicks "Cancel Name Request" button. */
-  async cancelPayment () {
-    // rollback the NR
-    this.$props.onCancel()
-    // close this modal
-    this.hideModal()
-  }
-
   @Watch('isVisible')
   onVisibleChanged (val: boolean) {
     if (val && this.$el?.querySelector) {
       this.$nextTick(() => {
         // add classname to button text (for more detail in Sentry breadcrumbs)
-        const confirmNrCancelBtn = this.$el.querySelector("#payment-cancel-btn > span")
-        if (confirmNrCancelBtn) confirmNrCancelBtn.classList.add("confirm-nr-cancel-btn")
-        const confirmNrContinueBtn = this.$el.querySelector("#payment-pay-btn > span")
-        if (confirmNrContinueBtn) confirmNrContinueBtn.classList.add("confirm-nr-continue-btn")
-        const confirmNrCloseBtn = this.$el.querySelector("#payment-close-btn > span")
-        if (confirmNrCloseBtn) confirmNrCloseBtn.classList.add("confirm-nr-close-btn")
+        const upgradeAcceptBtn = this.$el.querySelector("#payment-pay-btn > span")
+        if (upgradeAcceptBtn) upgradeAcceptBtn.classList.add("upgrade-accept-btn")
+        const upgradeCancelBtn = this.$el.querySelector("#payment-close-btn > span")
+        if (upgradeCancelBtn) upgradeCancelBtn.classList.add("upgrade-cancel-btn")
       })
     }
   }
