+++ src/views/landing.vue
@@ -2,49 +2,8 @@
   <v-container id="landing-container" class="ma-0 pa-0" fluid>
     <v-row id="upper-row" no-gutters align-content="start">
       <v-container class="landing-content-container">
-        <v-col cols="12" class="mb-4">
-          <Stats class="pr-10"/>
-        </v-col>
-        <v-col cols="12" class="mb-n2">
-          <span class="h2 colour-white">Name Request
-            <v-tooltip bottom nudge-right="10"
-                       nudge-top="5"
-                       content-class="bottom-tooltip tooltip-extra-wide"
-                       transition="fade-transition"
-                       class="test-class"
-                       :open-on-hover="false">
-              <template v-slot:activator="{ on }">
-                <v-btn class="beta-wrapper-btn ml-n2 mb-n2 pa-0"
-                       @click="on.click"
-                       @blur="on.blur"
-                       :ripple="false"
-                       retain-focus-on-click>
-                  <sup class="beta-tag">Beta</sup>
-                </v-btn>
-              </template>
-              <p>The Name Request website is available as a Beta version. Name requests obtained through the Name
-                Request Beta website are official name requests and can be used in the province of British Columbia.</p>
-              <p class="mb-0">The Name Request Beta website contains all of the functionality of the previous Name
-                Requests Online site, as well as new functionality intended to improve the experience of starting and
-                managing businesses in B.C. Some features of the site may not function as intended, and if you encounter
-                this please contact us:</p>
-              <span class="tooltip-bullet mt-3">
-                <a class="white--text" href="tel:+1-877-370-1033" target="_blank">Toll Free: 1-877-370-1033</a>
-              </span>
-              <span class="tooltip-bullet">
-                <a class="white--text" href="tel:+1-250-370-1033" target="_blank">Victoria Office: 250-370-1033</a>
-              </span>
-              <span class="tooltip-bullet mb-3">
-                <a class="white--text" href="mailto:BCRegistries@gov.bc.ca" target="_blank">BCRegistries@gov.bc.ca</a>
-              </span>
-              <p class="mb-0">As part of a
-                <a class="white--text" :href="agileUrl" target="_blank">modern agile software development process</a>
-                the Name Request website will be continually updated and improved based on feedback from citizens and
-                businesses in B.C.
-              </p>
-            </v-tooltip>
-          </span>
-        </v-col>
+        <AppTitleCols />
+
         <div class="main-container-style mt-3">
           <transition name="fade" mode="out-in" :duration="{ enter: 100, leave: 100 }">
             <keep-alive :include="['Tabs']">
@@ -54,6 +13,7 @@
         </div>
       </v-container>
     </v-row>
+
     <v-row id="lower-row" no-gutters>
       <LowerContainer />
     </v-row>
@@ -63,13 +23,15 @@
 <script lang="ts">
 import { Vue, Component, Prop } from 'vue-property-decorator'
 
+import { AppTitleCols } from '@/components/common'
+import LowerContainer from '@/components/lower-info-area/lower-container.vue'
+
 import AnalyzeCharacters from '@/components/new-request/analyze-characters.vue'
 import AnalyzePending from '@/components/new-request/analyze-pending.vue'
 import AnalyzeResults from '@/components/new-request/analyze-results.vue'
 import SendToExamination from '@/components/new-request/send-to-examination.vue'
 import ExistingRequestDisplay from '@/components/existing-request/existing-request-display.vue'
 import ExistingRequestEdit from '@/components/existing-request/existing-request-edit.vue'
-import LowerContainer from '@/components/lower-info-area/lower-container.vue'
 import QuickSearchPending from '@/components/new-request/quick-search-pending.vue'
 import QuickSearchResults from '@/components/new-request/quick-search-results.vue'
 import SearchPending from '@/components/existing-request/search-pending.vue'
@@ -81,6 +43,7 @@ import newRequestModule from '@/store/new-request-module'
 
 @Component({
   components: {
+    AppTitleCols,
     AnalyzeCharacters,
     AnalyzePending,
     AnalyzeResults,
@@ -122,41 +85,17 @@ export default class Landing extends Vue {
 </script>
 
 <style lang="scss" scoped>
-@import '@/assets/scss/theme.scss';
+/*@import '@/assets/scss/theme.scss';*/
 
 .landing-content-container {
   min-width: 940px;
   max-width: 1140px;
 }
+
 #upper-row {
   background: url('../assets/images/analyze-name-bg.jpg') no-repeat bottom;
   background-size: cover;
   color: white;
   min-height: 700px;
-}
-.beta-tag {
-  top: -1rem;
-  color: $BCgovGold5;
-  font-size: 19px;
-  font-weight: bold;
-  text-transform: uppercase;
-  &:hover {
-    cursor: pointer;
-  }
-}
-.beta-wrapper-btn {
-  background-color: inherit !important;
-  box-shadow: unset !important;
-  font: inherit !important;
-  height: inherit !important;
-  &:before {
-    background-color: inherit !important;
-  }
-}
-.v-tooltip__content {
-  pointer-events: auto !important;
-}
-.tooltip-extra-wide {
-  width: 410px !important;
 }
 </style>
